seperate react-redux log tools
