+++ frontend/react/containers/root.jsx
@@ -3,16 +3,19 @@ import { Provider } from 'react-redux';
 import { Route, Router, IndexRoute } from 'react-router';
 
 import CounterComponent from './../containers/counter-containers';
+import DevTools from './dev-tools';
 
 export default ({ store, history }) => {
   return (
     <Provider store={store}>
-      <Router history={history}>
-        <Route path='/' >
-          <IndexRoute component={CounterComponent} />
-          <Route path='about' component={CounterComponent2} />
-        </Route>
-      </Router>
+      <div>
+        <Router history={history}>
+          <Route path='/' >
+            <IndexRoute component={CounterComponent} />
+          </Route>
+        </Router>
+        <DevTools />
+      </div>
     </Provider>
   );
 };
