+++ frontend/react/store/configure-store.dev.js
@@ -1,0 +1,13 @@
+import { createStore, applyMiddleware } from 'redux';
+import reducers from './../reducers';
+import DevTools from './../containers/dev-tools';
+
+export default (initialState = {}, { reactRouterMiddleware }) => {
+  const store = createStore(
+    reducers,
+    initialState,
+    DevTools.instrument(),
+    applyMiddleware(reactRouterMiddleware),
+  );
+  return store;
+};
